ci/requirements.txt
jsonschema==4.22.0
pyyaml==6.0.2
openapi-spec-validator==0.7.1
